refactor(class Website): adds better access control
- renames old class Website into class Location
- makes a new class Website a thin wrapper around Location
- compiles but not tested
