+++ inc/Website.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "HttpMethods.hpp"
-#include "Token.hpp"
 #include <list>
 #include <map>
 #include <ostream>
@@ -15,92 +13,19 @@ struct Listen {
 };
 
 class PathInfo;
-class Website;
+class Location;
+class Token;
 
 class Website {
 public:
-  typedef enum {
-    None,
-    Cgi,
-    Return,
-    Redirect
-  } Type;
-
-  Website();
-  Website(const Website &Other);
-  Website &operator=(const Website &Other);
-  ~Website();
-
-  explicit Website(const std::string &Path);
-
-  bool getAutoindex(void) const;
-  const char *getErrorPage(const unsigned int Code) const;
-  unsigned int getMaxReqBody(void) const;
+  explicit Website(const Location &Loc);
+  ~Website(void);
   PathInfo getPathInfo(const std::string &Path) const;
-
-  int getAllow(void) const;
-  explicit Website(std::list<Token>::const_iterator It);
-  void addInterface(Listen &If);
-  /// \fun addLocation
-  /// \brief if a location with the same path already exists it is replaced
-  void addLocation(Website &Loc);
-  void addErrorPage(const unsigned int Code, const std::string &Path);
   const std::list<Listen> &getInterfaces(void) const;
-  const std::list<Website> &getLocations(void) const;
-  const std::map<unsigned int, std::string> &getErrorPages(void) const;
-  void setRoot(const std::string &RootDir);
-  void setAutoindex(const bool IsOn);
-  void setMaxReqBody(const unsigned int MaxBody);
-  void addAllow(const HttpMethod Method);
-  void allowAll(void);
-  void allowNone(void);
-  void setAsRoot(void);
-  const std::string &getRoot(void) const;
-
-  unsigned int getReturnCode(void) const;
-  const std::string &getReturnPath(void) const;
-
-  const std::string &getPath(void) const;
-  const std::string &getRedirect(void) const;
-  const std::string &getCgi(void) const;
-  Type getType(void) const;
-
-  void setReturn(const unsigned int Code, const::std::string &Url);
-  void setCgi(const::std::string &Path);
-  void setRedirect(const::std::string &Path);
-
-  void print(std::ostream &Os) const;
-
-  bool isSetAutoindex(void) const;
-  bool isSetRoot(void) const;
-  bool isSetInterfaces(void) const;
-  bool isSetAllow(void) const;
-  bool isSetMaxReqBody(void) const;
-  bool isRoot(void) const;
+  std::ostream &print(std::ostream &Os) const; 
 
 private:
-  typedef enum {
-    Interfaces = (1u << 0),
-    Root = (1u << 1),
-    Autoindex = (1u << 2),
-    Allow = (1u << 3),
-    MaxReqBody = (1u << 4)
-  } SetMembers;
-
-  bool _isRoot;
-  int _setMembers;
-  unsigned int _maxReqBody;
-  std::list<Listen> _interfaces; 
-  std::list<Website> _locations; 
-  std::string _root;
-  std::map<unsigned int, std::string> _errorPages;
-  bool _autoindex;
-  int _allow;
-  Type _type;
-  std::string _path;
-  std::string _redirect;
-  unsigned int _returnCode;
-  std::string _returnPath;
+  const Location &_website;
 };
 
 class PathInfo {
@@ -116,7 +41,7 @@ public:
   PathInfo &operator=(const PathInfo &Other);
   ~PathInfo(void);
 
-  PathInfo(const Website &Site, const std::string &Path);
+  PathInfo(const Location &Site, const std::string &Path);
 
   unsigned int getMaxReqBody(void) const;
   const std::string &getRoot(void) const;
@@ -130,8 +55,8 @@ public:
   void print(std::ostream &Os) const;
 
 private:
-  void resolveLocations(const std::list<Website> &Locations);
-  void populateFromLocation(const Website &Loc);
+  void resolveLocations(const std::list<Location> &Locations);
+  void populateFromLocation(const Location &Loc);
   std::string _cgiPath;
   std::string _realPath;
   Action _action;
