+++ src/config/Parser.hpp
@@ -1,6 +1,9 @@
 #pragma once
 
 #include "Config.hpp"
+#include "Scanner.hpp"
+#include "Token.hpp"
+#include "Website.hpp"
 
 class Parser {
 public:
@@ -32,31 +35,31 @@ private:
   bool isNextType(const TokenType::Type Type) const;
   const std::string &matchGetLexeme(TokenType::Type Type);
 
-  Website expression(void);
+  Location expression(void);
 
   Website server(void);
 
   unsigned int parseUnsignedInt(void);
   void gap(void);
   void populateInterface(Listen &Interface);
-  void parseListen(Website &Site);
-  void parseRoot(Website &Site);
-  void parseAutoindex(Website &Site);
-  void parseAllow(Website &Site);
+  void parseListen(Location &Site);
+  void parseRoot(Location &Site);
+  void parseAutoindex(Location &Site);
+  void parseAllow(Location &Site);
 
-  void parseLocation(Website &Site);
-  void parseRedirect(Website &Site);
-  void parseCgi(Website &Site);
-  void parseReturn(Website &Site);
-  void parseMaxReqBody(Website &Site);
+  void parseLocation(Location &Site);
+  void parseRedirect(Location &Site);
+  void parseCgi(Location &Site);
+  void parseReturn(Location &Site);
+  void parseMaxReqBody(Location &Site);
 
   void addIpv4(Listen &Interface);
   void addPort(Listen &Interface);
 
-  void parseErrorPage(Website &Site);
-  void validateEntry(const Website &Site);
+  void parseErrorPage(Location &Site);
+  void validateEntry(const Location &Site);
 
-  void parseEntry(Website &Website);
+  void parseEntry(Location &Website);
   std::string parseResource(void);
   std::string parseAbsPath(void);
   std::string parseWord(void);
